+++ packages/gui/src/lib/codegen/emotion.ts
@@ -2,11 +2,12 @@ import { toH } from 'hast-to-hyperscript'
 import { HtmlNode } from '../../components/html/types'
 import { editorSchemaToHast } from '../transformers/editor-schema-to-hast'
 import { toCSSObject } from './to-css-object'
-import { stringifyHastNode } from './stringify-hast-node'
+import { stringifyHastNode } from './stringify-hast-node-as-jsx'
 import { toReactProps } from './to-react-props'
 import { format } from './format'
 import { getPropSyntax } from './util'
 import { Theme } from '../../types/theme'
+import { CodegenOptions } from './types'
 
 const h =
   (theme?: Theme) => (tagName: string, props: any, children?: any[]) => {
@@ -21,9 +22,9 @@ const h =
     return { tagName, props: newProps, children }
   }
 
-export const emotion = async (node: HtmlNode, options: any) => {
+export const emotion = async (node: HtmlNode, options: CodegenOptions) => {
   const root = editorSchemaToHast(node, { addSlotSyntax: true })
-  const functionBody = stringifyHastNode(toH(h(options.theme), root))
+  const functionBody = stringifyHastNode(toH(h(options?.theme), root))
 
   const output = `
   /** @jsxImportSource @emotion/react */