+++ packages/gui/src/lib/codegen/react.ts
@@ -3,8 +3,9 @@ import { HtmlNode } from '../../components/html/types'
 import { extractStyles } from './extract-styles'
 import { format } from './format'
 import { html as toHtml } from './html'
-import { stringifyHastNode } from './stringify-hast-node'
+import { stringifyHastNode } from './stringify-hast-node-as-jsx'
 import { toReactProps } from './to-react-props'
+import { CodegenOptions } from './types'
 
 const h = (tagName: string, props: any, children?: any[]) => {
   const newProps = toReactProps(props)
@@ -12,8 +13,11 @@ const h = (tagName: string, props: any, children?: any[]) => {
   return { tagName, props: newProps, children }
 }
 
-export const react = async (node: HtmlNode): Promise<string> => {
-  const html = await toHtml(node)
+export const react = async (
+  node: HtmlNode,
+  options: CodegenOptions
+): Promise<string> => {
+  const html = await toHtml(node, options)
   const { styles } = await extractStyles(html)
   const jsx = stringifyHastNode(toH(h, node as any))