+++ packages/gui/src/lib/codegen/theme-ui.ts
@@ -2,26 +2,32 @@ import { toH } from 'hast-to-hyperscript'
 import { HtmlNode } from '../../components/html/types'
 import { editorSchemaToHast } from '../transformers/editor-schema-to-hast'
 import { toCSSObject } from './to-css-object'
-import { stringifyHastNode } from './stringify-hast-node'
+import { stringifyHastNode } from './stringify-hast-node-as-jsx'
 import { toReactProps } from './to-react-props'
 import { format } from './format'
 import { getPropSyntax } from './util'
+import { CodegenOptions } from './types'
+import { Theme } from '../../types/theme'
 
-const h = (tagName: string, props: any, children?: any[]) => {
-  const newProps = toReactProps(props)
+const h =
+  (theme?: Theme) => (tagName: string, props: any, children?: any[]) => {
+    const newProps = toReactProps(props)
 
-  if (newProps.style) {
-    const style = newProps.style
-    delete newProps.style
-    newProps.sx = toCSSObject(style)
+    if (newProps.style) {
+      const style = newProps.style
+      delete newProps.style
+      newProps.sx = toCSSObject(style, theme)
+    }
+
+    return { tagName, props: newProps, children }
   }
 
-  return { tagName, props: newProps, children }
-}
-
-export const themeUI = async (node: HtmlNode) => {
+export const themeUI = async (
+  node: HtmlNode,
+  { theme }: CodegenOptions = {}
+) => {
   const root = editorSchemaToHast(node, { addSlotSyntax: true })
-  const functionBody = stringifyHastNode(toH(h, root))
+  const functionBody = stringifyHastNode(toH(h(theme), root))
 
   const output = `
   /** @jsxImportSource theme-ui */