+++ packages/gui/src/lib/transformers/plugins/convert-components-to-hast.ts
@@ -4,5 +4,9 @@ export const convertComponentsToHast = () => (tree: any) => {
   visit(tree, 'component', (node) => {
     node.type = 'element'
     node.component = true
+    node.attributes = {
+      ...node.attributes,
+      ...node.value.attributes,
+    }
   })
 }